Day - 2: Bogged Down By School
Been super busy with school as next week is finals week.

Used another sliding window solution. I was able to simplify the problem
a lot by realizing that all we care about is the max of each color in
the game instance. Once we know this it just becomes a simple sliding
window problem, and we check the current num/color against the max.

Also did not have to do any crazy debugging, my solution was accepted
first try in part-1 and part-2.