Day - 6
This was a pretty fun one. Much easier than Day 5. Could only think of a
brute forced solution, but I was able to find an optimization were we
only find the upper and lower bound of the valid range of charge times.
Then, we can find the number of valid solutions by subtracting the upper
bound from the lower bound and adding 1.