Day - 8
Part 1 was very easy and self explanitory. Also, came up with a solution
for part 2 that was very easy and straight forward, but the execution
time was too long. The "optimized" solution is to find the LCM of all
the end points for each starting point. This LCM is then the answer.
Absolutely no way I would have reached this solution on my own, but I
guess that it makes sense. All of the cycles have to reach each other
eventually, and the LCM finds that intersection. However, it makes a
large assumtion that a start will cycle from start to end at the same
fixed interval, which there is no mention of in the puzzle.