Day - 9
Tried my hand at a recursive function, which I don't know if it was
ideal, but was good practice. For part 2 I attempted to modify my
solution from part 1, which worked with the test inputs, but for some
reason failed on the full input. Instead, I just reversed the array and
passed it into the original function to solve part 2.